small refactoring for readability +fix gamebundle format
